zypper: don't ignore errors when getting the package name
This leads bad error messages with empty package names. Instead check
stderr. We can't check RC since this is the number of processed
packages.

In order to not trip on warnings ignore RPM signature validation errors
since they are of little interest when getting the package name and so
far we ignored all errors.